refactor(transform): isolate mutation analysis
